rename kernel host_utils to common_utils
