+++ ttnn/cpp/ttnn/kernel/kernel_common_utils.hpp
