+++ ttnn/cpp/ttnn/operations/experimental/where/device/kernels/dataflow/elemwise_reader_kernel_args.hpp
@@ -5,7 +5,7 @@
 
 #include <cstdint>
 
-#include "ttnn/kernel/kernel_host_utils.hpp"
+#include "ttnn/kernel/kernel_common_utils.hpp"
 
 namespace ttnn::kernel::eltwise::where_args {
